feat(connlib): allow and route packets to Gateway TUN IPs (#8294)
At the moment, `connlib` doesn't allow routing packets directly to
Gateways because the subnet we've chosen for the tunnel IPs isn't part
of the routing table. In addition, all traffic within `connlib` is
expected to be targeting a resource _beyond_ a Gateway.

In order to resolve SRV and TXT records within a certain site, we've
opted to host a DNS server on the Gateway's TUN device. See #8285 for
details on that. To actually reach that DNS server, we need to add a few
new control flows to `connlib` where we detect whether a packet is
directly for the tunnel IP of a Gateway or for a resource.

We only know a Gateway's IP once we are connected to it, meaning we
cannot route those packets prior to that. We also cannot establish a
connection when the user attempts to as every connection intent sent to
the portal needs to reference a Resource. For the usecase of resolving
SRV and TXT records, the packets will be associated with the DNS
resource for which we are trying to resolve records.

This patch only established the base connectivity and necessary
exceptions to the Client's filter rules in order to route packets to the
Gateway's TUN device. The following commands have been issued against a
staging Gateway, demonstrating connectivity to the Gateway's TUN device
from a Client after establishing a connection to it:

```
❯ ping github.com -c 1
PING github.com (fd00:2021:1111:8000::) 56 data bytes
64 bytes from github.com (fd00:2021:1111:8000::): icmp_seq=1 ttl=50 time=1441 ms

--- github.com ping statistics ---
1 packets transmitted, 1 received, 0% packet loss, time 0ms
rtt min/avg/max/mdev = 1440.614/1440.614/1440.614/0.000 ms

❯ ping 100.72.145.83 -c 1
PING 100.72.145.83 (100.72.145.83) 56(84) bytes of data.
64 bytes from 100.72.145.83: icmp_seq=1 ttl=64 time=213 ms

--- 100.72.145.83 ping statistics ---
1 packets transmitted, 1 received, 0% packet loss, time 0ms
rtt min/avg/max/mdev = 212.574/212.574/212.574/0.000 ms
```

Related: #8221