refactor(connlib): decouple mangled DNS queries from DNS mapping (#8331)
When `connlib` receives a UDP packet for one of its DNS resolver IPs and
determines that it needs to be forwarded to another resolver through the
tunnel, it mangles the destination IP + port to point to this new
resolver. In order for the response to be correctly recognised by the
application, the response packet needs its _source_ IP + port mangled.
This information is currently stored in a `HashMap` together with an
expiry timestamp.

To be precise, the information that is captured is only the new
destination socket, not the current one. The old socket is then later
implied by the DNS mapping that we remember internally, i.e. which one
of `connlib`'s DNS resolver IPs maps to which upstream DNS server.

For the usecase of forwarding DNS queries of type SRV and TXT to the
site that hosts the DNS resource in question, we want to send those DNS
queries to a Gateway within that site. For UDP DNS queries, this
requires the same data structure as we do for DNS queries that are
tunneled to another DNS resolver _beyond_ the Gateway. In fact, from the
perspective of the Client, there is no difference between a packet that
is handled by the Gateway or by a resolver behind the Gateway. The only
difference is in the new destination IP + port.

In the case where the Gateway is targeted with the DNS query, we won't
be able to resolve the original destination socket from the DNS mapping
data structure because the Gateway's IP isn't explicitly configured as a
DNS resolver.

To handle both of these cases with the same data structure, we refactor
this temporary mapping to simply store the original destination socket.
To make the data structure less complicated to use, we introduce an
`ExpiringMap` that automatically removes entries after a certain
deadline. This is important for UDP DNS queries to ensure this map
doesn't in an unbounded manner if for some reason, the configured DNS
resolver never replies.

Related: #8221