fix(connlib): Update search_domain for exsiting TunConfigs (#8445)
For existing `TunConfig`, we had a bug where we failed to update the
search_domain if the effective dns_servers were unchanged.

@thomaseizinger I can see why you want to refactor this; it's quite a
mess to follow ;-). I was going to try my hand at cleaning it up a
little bit just so I can grok it but I figured since this area is going
to be changing quite a bit in #8263, I'll leave those changes out for
now.