fix(connlib): remove explicit `Session::disconnect`  (#8474)
Within the event-loop, we already react to the channel being closed
which happens when the `Sender` within the `Session` gets dropped. As
such, there is no need to send an explicit `Stop` command, dropping the
`Session` is equivalent.

As it turns out, `swift-bridge` already calls `Drop` for us when the
last pointer is set to `nil`:
https://github.com/firezone/firezone/blob/280a9dd999e4754ac57c768b1d423b8a358cfe40/swift/apple/FirezoneNetworkExtension/Connlib/Generated/connlib-client-apple/connlib-client-apple.swift#L24-L28

Thus, we can also remove the explicit `disconnect` call to
`WrappedSession` entirely.