ci(rust): install nightly toolchain (#8507)
For #8501, we need to install a nightly toolchain in our CI system in
order to compile to eBPF kernel. We already use a nightly toolchain for
one of the static analysis tools.

In this PR, we extend our `setup-rust` action to install the nightly
toolchain for us which allows us to reuse that later.